# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIXTYSIX API",
    "version": "0.2.0",
    "description": "구현 전 계약. B0 로그인 2종/공개 모집 반영. 서버 시간 권위, unknown request fields 거부, 업무 성공 쓰기만 멱등 캐시. 세부 권장 정책은 BACKEND-B0-SPEC.md 참조. 인증 cookie 실제 이름은 T03 adapter에서 고정. 모든 보호/인증 응답 Cache-Control: no-store."
  },
  "servers": [
    {
      "url": "/",
      "description": "동일 origin /v1 프록시; 배포 미구현"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/v1/auth/kakao/start": {
      "get": {
        "operationId": "kakaoStart",
        "summary": "카카오 인증 시작",
        "tags": [
          "auth"
        ],
        "responses": {
          "302": {
            "description": "성공",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri"
                },
                "description": "허용된 Kakao authorize URL"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "intentId 없으면 로그인. 있으면 현재 세션의 link/reauth intent만 허용. 목적·브라우저에 결합된 state를 생성. state 저장 실패 시 redirect하지 않음. state hash·5분 만료·S256 PKCE 적용. reauth에는 prompt=login. 알려진 카카오 인앱 브라우저의 reauth 시작은 403. intent당 state 하나만 발급하며 재시작은 새 intent로 진행.",
        "parameters": [
          {
            "in": "query",
            "name": "returnTo",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "/home",
                "/my",
                "/onboarding"
              ]
            }
          },
          {
            "in": "query",
            "name": "intentId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": []
      }
    },
    "/v1/auth/kakao/callback": {
      "get": {
        "operationId": "kakaoCallback",
        "summary": "카카오 인증 callback",
        "tags": [
          "auth"
        ],
        "responses": {
          "302": {
            "description": "성공",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "allowlist 서비스 화면. 새 세션은 로그인 성공에만 설정"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "code 또는 error 중 하나. state 일회성·목적·브라우저·intent 소유권 검사 후 서버 token 교환. providerSubject로 조회, 이메일만으로 병합하지 않음. 실패는 고정 서비스 화면의 공개 오류 코드로 redirect. 원문 provider 오류/token을 URL에 전달하지 않음.",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "query",
            "name": "error_description",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 2048
            },
            "description": "공급자 호환성용 입력. 저장·로그·반환하지 않음."
          }
        ],
        "security": []
      }
    },
...
    "/v1/me/reauth-intents": {
      "post": {
        "operationId": "createReauthIntent",
        "summary": "기존 로그인 수단으로 재인증 시작",
        "tags": [
          "me"
        ],
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Intent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "본인 소유의 사용 가능한 identity만 허용. callback/OTP 완료는 동일 user+session만 갱신. 다른 계정 로그인으로 교체하지 않음.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "identityId": {
                    "type": "string",
                    "minLength": 1,
                    "format": "uuid"
                  }
                },
                "required": [
                  "identityId"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/me/identities/link-intents": {
      "post": {
        "operationId": "createLinkIntent",
        "summary": "로그인 수단 연결 시작",
        "tags": [
          "me"
        ],
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Intent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "최근 5분 내 기존 수단 재인증 필수. 이메일이면 OTP challenge, 카카오면 redirect URL. 타 계정에 연결된 수단은 연결 완료 때 409.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "provider": {
                        "type": "string",
                        "enum": [
                          "email"
                        ]
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 254
                      }
                    },
                    "required": [
                      "provider",
                      "email"
                    ]
                  },
                  {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "provider": {
                        "type": "string",
                        "enum": [
                          "kakao"
                        ]
                      }
                    },
                    "required": [
                      "provider"
                    ]
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/v1/me/identities/link-intents/{intentId}/complete": {
      "post": {
        "operationId": "completeLink",
        "summary": "검증 완료된 수단 연결",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Identity"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "현재 user/session에 결합된 만료 전 서버 검증 intent를 원자적으로 소비. 클라이언트 verified 플래그나 providerSubject를 받지 않음. 기존 계정·기록 합병 없음.",
        "parameters": [
          {
            "in": "path",
            "name": "intentId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {}
              }
            }
          }
        }
      }
    },
    "/v1/me/identities/{identityId}": {
      "delete": {
        "operationId": "unlinkIdentity",
        "summary": "로그인 수단 해제",
        "tags": [
          "me"
        ],
        "responses": {
          "204": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "최근 5분 재인증, 사용자 잠금, 다른 사용 가능한 로그인 수단 존재 확인. 라이브러리 account 행 개수만 세지 않음.",
        "parameters": [
          {
            "in": "path",
            "name": "identityId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "summary": "내 계정·현재 참여",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-sixtysix.session",
        "description": "HTTPS: __Host-sixtysix.session, Path=/·HttpOnly·Secure·SameSite=Lax·Domain 없음. 로컬 HTTP만 sixtysix.session. DB에는 토큰 SHA-256만 저장, 개발 기본 절대 만료 7일."
      }
    },
    "parameters": {
...
      "Csrf": {
        "in": "header",
        "name": "X-CSRF-Token",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "GET /v1/auth/context 발급값. 브라우저/세션에 결합, Origin 검사와 함께 사용."
      },
...
    },
    "responses": {
      "400": {
        "description": "입력 또는 인증 challenge/state 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "401": {
        "description": "유효 세션 없음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "403": {
        "description": "운영 권한 부족 또는 REAUTH_REQUIRED",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "404": {
        "description": "없거나 접근 불가한 업무 리소스",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "409": {
        "description": "업무/멱등/연결 충돌",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "429": {
        "description": "속도 제한",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "재시도까지 초"
          }
        }
      },
      "503": {
        "description": "일시 장애",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Identity": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "provider": {
            "type": "string",
            "enum": [
              "email",
              "kakao"
            ]
          },
          "label": {
            "type": "string",
            "description": "본인에게만 보이는 마스킹된 식별자"
          },
          "usable": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "provider",
          "label",
          "usable"
        ]
      },
      "Profile": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          }
        },
        "required": [
          "id",
          "displayName"
        ]
      },
      "Preferences": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "defaultVisibility": {
            "type": "string",
            "enum": [
              "cohort",
              "private"
            ]
          }
        },
        "required": [
          "defaultVisibility"
        ]
      },
      "Membership": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cohortId": {
            "type": "string",
            "format": "uuid"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "leftAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "cohortId",
          "joinedAt",
          "cancelledAt",
          "leftAt"
        ]
      },
      "Me": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "identities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Identity"
            }
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Membership"
            }
          },
          "currentMembershipId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "profile",
          "identities",
          "preferences",
          "memberships",
          "currentMembershipId"
        ]
      },
...
      "Available": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "targetDay": {
            "type": "integer",
            "minimum": 1,
            "maximum": 66
          },
          "targetDate": {
            "type": "string",
            "format": "date"
          },
          "late": {
            "type": "boolean"
          },
          "closesAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "ok",
          "targetDay",
          "targetDate",
          "late",
          "closesAt"
        ]
      },
      "Unavailable": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "BEFORE_START",
              "ALREADY_FILLED",
              "ENDED",
              "CANCELLED",
              "LEFT"
            ]
          }
        },
        "required": [
          "ok",
          "reason"
        ]
      },
      "Availability": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/Available"
          },
          {
            "$ref": "#/components/schemas/Unavailable"
          }
        ]
      },
...
      "Intent": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "enum": [
              "email",
              "kakao"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "nextAction": {
            "type": "string",
            "enum": [
              "ENTER_CODE",
              "REDIRECT"
            ]
          },
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "authorizationUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          }
        },
        "required": [
          "id",
          "provider",
          "expiresAt",
          "nextAction",
          "challengeId",
          "authorizationUrl"
        ]
      },
...
      "ErrorDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "field": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INVALID_REQUEST",
                  "INVALID_TEXT",
                  "INVALID_MEDIA",
                  "INVALID_CHALLENGE",
                  "INVALID_OAUTH_STATE",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "REAUTH_REQUIRED",
                  "NOT_FOUND",
                  "COHORT_FULL",
                  "JOIN_CLOSED",
                  "ACTIVE_MEMBERSHIP_EXISTS",
                  "ALREADY_FILLED",
                  "PASS_EXHAUSTED",
                  "BEFORE_START",
                  "ENDED",
                  "TARGET_CHANGED",
                  "IDEMPOTENCY_CONFLICT",
                  "IDENTITY_ALREADY_LINKED",
                  "ACCOUNT_NOT_LINKED",
                  "LAST_IDENTITY",
                  "INVALID_STATE",
                  "RATE_LIMITED",
                  "TEMPORARILY_UNAVAILABLE",
                  "COHORT_EXISTS"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "requestId": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "$ref": "#/components/schemas/ErrorDetails"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}